add get metrics for flyteRemote and add helper method explain/dump
